Add a service for testing the errors modules and moved the ApolloError class to services/util.py
